refresh image from first pixels instead of 1% of image

+++ fraktal_sft/fraktal_sft.cpp
@@ -1490,7 +1490,7 @@ void CFraktalSFT::Zoom(int nXPos, int nYPos, double nZoomSize, int nWidth, int n
 		RenderFractal(m_nX, m_nY, m_nMaxIter, m_hWnd);
 }
 
-int CFraktalSFT::GetProgress(int *pnGuessed, int *pnRDone, int *pnAP, int *pnT)
+double CFraktalSFT::GetProgress(int *pnGuessed, int *pnRDone, int *pnAP, int *pnT)
 {
 	if (pnGuessed)
 		*pnGuessed = m_nGuessed * 100.0 / (m_nDone ? m_nDone : 1);
